fix links when running straight from files
When you run this in node.js or from the website, everything is fine. But if you run it straight from the files app on your device the links to each game doesent work. This fixes that by going straight to the file. I dont know If this breaks running it on the website or whatever but I dont see why it would. :)
I dont know much about javascript or web development, but theres probably some fancy javascript way to tell if its being run on localhost or whatever or just from the files app.